rename models and panel modules for clarity
